Added door and laser-receiver - Optimized other assets
Door model with blueprint functionality has been added, along with a
laser receiver model and blueprint.

The laser receiver now utilizes an interface IActivateable to activate
assigned objects that implements this interface. For now, the door is
the only blueprint implementing this.

The walls have been updated to save RAM, and also perform better with
baked lighting.