Updated the Character
Added the correct model and animations to the character.

Made the character able to respond both to gamepad and mouse/keyboard
controls.

Made a MouseCursor, which visualizes the current position of the mouse
when the mouse is being used to control the character.

Updated glass shader.

Updated level blueprint to implement the ILevel interface, making it
easy to interact with the level blueprint from other blueprint classes.