Updated a couple of blueprints
Updated: Vortexfield now always turns the laser 90 degrees, even when
hit on the edge.
Updated: LaserBase now includes a property for an FMOD event, to play a
sound on explosion.
Added: Particle effect for when the laser is exploding
Updated: Floor1new, now utilizes a simple floor blueprint instead of
planes, this needs to be updated in floor 0 aswell.